+++ plexar/core/service.py
@@ -12,200 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import asyncio
-import uuid
 from logging import getLogger
-from typing import Callable, Dict, List, Literal, Optional, Tuple, Union
+from typing import Callable, Dict, List, Optional, Tuple
 
 import xoscar as xo
-from fastapi import APIRouter, FastAPI, Request
-from fastapi.middleware.cors import CORSMiddleware
-from fastapi.responses import JSONResponse
-from pydantic import BaseModel, Field
-from typing_extensions import NotRequired, TypedDict
-from uvicorn import Config, Server
 
 from plexar.core import ModelActor
 from plexar.model import ModelSpec
-from plexar.model.llm.types import ChatCompletion, Completion
-
-max_tokens_field = Field(
-    default=16, ge=1, le=2048, description="The maximum number of tokens to generate."
-)
-
-temperature_field = Field(
-    default=0.8,
-    ge=0.0,
-    le=2.0,
-    description="Adjust the randomness of the generated text.\n\n"
-    + "Temperature is a hyperparameter that controls the randomness of the generated text. It affects the probability distribution of the model's output tokens. A higher temperature (e.g., 1.5) makes the output more random and creative, while a lower temperature (e.g., 0.5) makes the output more focused, deterministic, and conservative. The default value is 0.8, which provides a balance between randomness and determinism. At the extreme, a temperature of 0 will always pick the most likely next token, leading to identical outputs in each run.",
-)
-
-top_p_field = Field(
-    default=0.95,
-    ge=0.0,
-    le=1.0,
-    description="Limit the next token selection to a subset of tokens with a cumulative probability above a threshold P.\n\n"
-    + "Top-p sampling, also known as nucleus sampling, is another text generation method that selects the next token from a subset of tokens that together have a cumulative probability of at least p. This method provides a balance between diversity and quality by considering both the probabilities of tokens and the number of tokens to sample from. A higher value for top_p (e.g., 0.95) will lead to more diverse text, while a lower value (e.g., 0.5) will generate more focused and conservative text.",
-)
-
-stop_field = Field(
-    default=None,
-    description="A list of tokens at which to stop generation. If None, no stop tokens are used.",
-)
-
-stream_field = Field(
-    default=False,
-    description="Whether to stream the results as they are generated. Useful for chatbots.",
-)
-
-top_k_field = Field(
-    default=40,
-    ge=0,
-    description="Limit the next token selection to the K most probable tokens.\n\n"
-    + "Top-k sampling is a text generation method that selects the next token only from the top k most likely tokens predicted by the model. It helps reduce the risk of generating low-probability or nonsensical tokens, but it may also limit the diversity of the output. A higher value for top_k (e.g., 100) will consider more tokens and lead to more diverse text, while a lower value (e.g., 10) will focus on the most probable tokens and generate more conservative text.",
-)
-
-repeat_penalty_field = Field(
-    default=1.1,
-    ge=0.0,
-    description="A penalty applied to each token that is already generated. This helps prevent the model from repeating itself.\n\n"
-    + "Repeat penalty is a hyperparameter used to penalize the repetition of token sequences during text generation. It helps prevent the model from generating repetitive or monotonous text. A higher value (e.g., 1.5) will penalize repetitions more strongly, while a lower value (e.g., 0.9) will be more lenient.",
-)
-
-presence_penalty_field = Field(
-    default=0.0,
-    ge=-2.0,
-    le=2.0,
-    description="Positive values penalize new tokens based on whether they appear in the text so far, increasing the model's likelihood to talk about new topics.",
-)
-
-frequency_penalty_field = Field(
-    default=0.0,
-    ge=-2.0,
-    le=2.0,
-    description="Positive values penalize new tokens based on their existing frequency in the text so far, decreasing the model's likelihood to repeat the same line verbatim.",
-)
-
-mirostat_mode_field = Field(
-    default=0,
-    ge=0,
-    le=2,
-    description="Enable Mirostat constant-perplexity algorithm of the specified version (1 or 2; 0 = disabled)",
-)
-
-mirostat_tau_field = Field(
-    default=5.0,
-    ge=0.0,
-    le=10.0,
-    description="Mirostat target entropy, i.e. the target perplexity - lower values produce focused and coherent text, larger values produce more diverse and less coherent text",
-)
-
-mirostat_eta_field = Field(
-    default=0.1, ge=0.001, le=1.0, description="Mirostat learning rate"
-)
-
-
-# CreateComplete request and response
-class CreateCompletionRequest(BaseModel):
-    prompt: str
-    suffix: Optional[str] = Field(None)
-    max_tokens: int = max_tokens_field
-    temperature: float = temperature_field
-    top_p: float = top_p_field
-    mirostat_mode: int = mirostat_mode_field
-    mirostat_tau: float = mirostat_tau_field
-    mirostat_eta: float = mirostat_eta_field
-    echo: bool = Field(
-        default=False,
-        description="Whether to echo the prompt in the generated text. Useful for chatbots.",
-    )
-    stop: Optional[Union[str, List[str]]] = stop_field
-    stream: bool = stream_field
-    logprobs: Optional[int] = Field(
-        default=None,
-        ge=0,
-        description="The number of logprobs to generate. If None, no logprobs are generated.",
-    )
-    presence_penalty: Optional[float] = presence_penalty_field
-    frequency_penalty: Optional[float] = frequency_penalty_field
-    logit_bias: Optional[Dict[str, float]] = Field(None)
-
-    model: str
-    n: Optional[int] = 1
-    best_of: Optional[int] = 1
-    user: Optional[str] = Field(None)
-
-    # llama.cpp specific parameters
-    top_k: int = top_k_field
-    repeat_penalty: float = repeat_penalty_field
-    logit_bias_type: Optional[Literal["input_ids", "tokens"]] = Field(None)
-
-    class Config:
-        schema_extra = {
-            "example": {
-                "prompt": "\n\n### Instructions:\nWhat is the capital of France?\n\n### Response:\n",
-                "stop": ["\n", "###"],
-            }
-        }
-
-
-# TODO: create embedding request and response
-class CreateEmbeddingRequest(BaseModel):
-    model: str
-    input: Union[str, List[str]] = Field(description="The input to embed.")
-    user: Optional[str]
-
-    class Config:
-        schema_extra = {
-            "example": {
-                "input": "The food was delicious and the waiter...",
-            }
-        }
-
-
-class ChatCompletionRequestMessage(TypedDict):
-    role: Literal["assistant", "user", "system"]
-    content: str
-    user: NotRequired[str]
-
-
-class CreateChatCompletionRequest(BaseModel):
-    messages: List[ChatCompletionRequestMessage] = Field(
-        default=[], description="A list of messages to generate completions for."
-    )
-    max_tokens: int = max_tokens_field
-    temperature: float = temperature_field
-    top_p: float = top_p_field
-    mirostat_mode: int = mirostat_mode_field
-    mirostat_tau: float = mirostat_tau_field
-    mirostat_eta: float = mirostat_eta_field
-    stop: Optional[List[str]] = stop_field
-    stream: bool = stream_field
-    presence_penalty: Optional[float] = presence_penalty_field
-    frequency_penalty: Optional[float] = frequency_penalty_field
-    logit_bias: Optional[Dict[str, float]] = Field(None)
-
-    model: str
-    n: Optional[int] = 1
-    user: Optional[str] = Field(None)
-
-    # llama.cpp specific parameters
-    top_k: int = top_k_field
-    repeat_penalty: float = repeat_penalty_field
-    logit_bias_type: Optional[Literal["input_ids", "tokens"]] = Field(None)
-
-    class Config:
-        schema_extra = {
-            "example": {
-                "messages": [
-                    {"role": "system", "content": "you are a helpful AI assistant"},
-                    {"role": "user", "content": "Hello!"},
-                    {"role": "assistant", "content": "Hi what can I help you?"},
-                ]
-            }
-        }
-
 
 logger = getLogger(__name__)
 
@@ -254,11 +67,6 @@ class SupervisorActor(xo.Actor):
             return target_worker
 
         raise RuntimeError("TODO")
-
-    async def get_worker(self, model_uid: str) -> xo.ActorRefType["WorkerActor"]:
-        if model_uid not in self._model_uid_to_worker:
-            raise Exception(f"Worker for model UID '{model_uid}' not found.")
-        return self._model_uid_to_worker[model_uid]
 
     @log
     async def launch_builtin_model(
@@ -313,156 +121,6 @@ class SupervisorActor(xo.Actor):
 
         worker_ref = await xo.actor_ref(address=worker_address, uid=WorkerActor.uid())
         self._worker_address_to_worker[worker_address] = worker_ref
-
-
-class RESTfulAPIActor(xo.Actor):
-    def __init__(self, host: str, port: int):
-        super().__init__()
-        self._supervisor_ref: xo.ActorRefType["SupervisorActor"]
-        app = FastAPI()
-        app.add_middleware(
-            CORSMiddleware,
-            allow_origins=["*"],
-            allow_credentials=True,
-            allow_methods=["*"],
-            allow_headers=["*"],
-        )
-        self.router = APIRouter()
-        self.router.add_api_route("/v1/models", self.list_models, methods=["GET"])
-        self.router.add_api_route("/v1/models", self.launch_model, methods=["POST"])
-        self.router.add_api_route(
-            "/v1/models/{model_uid}", self.terminate_model, methods=["DELETE"]
-        )
-        self.router.add_api_route(
-            "/v1/completions",
-            self.create_completion,
-            methods=["POST"],
-            response_model=Completion,
-        )
-        self.router.add_api_route(
-            "/v1/embeddings", self.create_embedding, methods=["POST"]
-        )
-        self.router.add_api_route(
-            "/v1/chat/completions",
-            self.create_chat_completion,
-            methods=["POST"],
-            response_model=ChatCompletion,
-        )
-        app.include_router(self.router)
-
-        # uvicorn
-        loop = asyncio.get_event_loop()
-        config = Config(app=app, loop=loop, host=host, port=port)
-        server = Server(config)
-        loop.create_task(server.serve())
-
-    @classmethod
-    def uid(cls) -> str:
-        return "plexar_RESTfulAPI"
-
-    async def __post_create__(self):
-        self._supervisor_ref = await xo.actor_ref(
-            address=self.address, uid=SupervisorActor.uid()
-        )
-
-    def gen_model_uid(self) -> str:
-        # generate a time-based uuid.
-        return str(uuid.uuid1())
-
-    async def list_models(self) -> List[str]:
-        models = await self._supervisor_ref.list_models()
-        return [model_uid for model_uid, _ in models]
-
-    async def launch_model(self, request: Request) -> str:
-        payload = await request.json()
-        model_name = payload.get("model_name")
-        model_size_in_billions = payload.get("model_size_in_billions")
-        model_format = payload.get("model_format")
-        quantization = payload.get("quantization")
-        kwargs = payload.get("kwargs", {}) or {}
-
-        model_uid = self.gen_model_uid()
-
-        await self._supervisor_ref.launch_builtin_model(
-            model_uid=model_uid,
-            model_name=model_name,
-            model_size_in_billions=model_size_in_billions,
-            model_format=model_format,
-            quantization=quantization,
-            **kwargs,
-        )
-        return JSONResponse(content={"model_uid": model_uid})
-
-    async def terminate_model(self, model_uid: str):
-        await self._supervisor_ref.terminate_model(model_uid)
-        return {"message": "Model terminated successfully."}
-
-    async def create_completion(self, request: Request, body: CreateCompletionRequest):
-        exclude = {
-            "prompt",
-            "model",
-            "n",
-            "best_of",
-            "logit_bias",
-            "logit_bias_type",
-            "user",
-        }
-        kwargs = body.dict(exclude=exclude)
-
-        if body.logit_bias is not None:
-            raise NotImplementedError
-        model_uid = body.model
-        worker_ref = await self._supervisor_ref.get_worker(model_uid)
-        model = await worker_ref.get_model(model_uid)
-
-        if body.stream:
-            raise NotImplementedError
-        else:
-            return await model.generate(body.prompt, kwargs)
-
-    async def create_embedding(self, request: CreateEmbeddingRequest):
-        raise NotImplementedError
-
-    async def create_chat_completion(
-        self,
-        request: Request,
-        body: CreateChatCompletionRequest,
-    ):
-        exclude = {
-            "n",
-            "model",
-            "messages",
-            "logit_bias",
-            "logit_bias_type",
-            "user",
-        }
-        kwargs = body.dict(exclude=exclude)
-
-        if body.logit_bias is not None:
-            raise NotImplementedError
-
-        user_messages = [
-            msg["content"] for msg in body.messages if msg["role"] == "user"
-        ]
-        if user_messages:
-            prompt = user_messages[-1]
-        else:
-            raise Exception("no prompt given")
-        system_prompt = next(
-            (msg["content"] for msg in body.messages if msg["role"] == "system"), None
-        )
-
-        chat_history = body.messages
-
-        model_uid = body.model
-        worker_ref = await self._supervisor_ref.get_worker(model_uid)
-        model = await worker_ref.get_model(model_uid)
-
-        if body.stream:
-            raise NotImplementedError
-
-        else:
-            return await model.chat(prompt, system_prompt, chat_history, kwargs)
 
 
 class WorkerActor(xo.Actor):
